Audit complet : fix crash simulate(), suppression log intrusif, +8 tests, doc
- Fix panic index out of range dans simulate() sur State malformé
  (Undecided + PossibleMoves vide) → retourne Stalemate
- Suppression log.Printf incontrôlable dans expandAll() (pollue benchmarks)
- Ajout type assertions défensives dans reasoning.Evaluator.Evaluate
- 8 nouveaux tests : fallback simulate, malformed state, PUCT négatif,
  expandAll vide, selectChildUCB AlphaZero, evaluator fallbacks
- Coverage mcts 94.6% → 98.6%, reasoning 92.9% → 98.2%
- Doc : traduction godoc AlphaZero en anglais, refs croisées backpropagate,
  documentation conversion valeur [0,1]→[-1,1]

+++ mcts/ucb1.go
@@ -25,7 +25,7 @@ func (n *mctsNode) ucb1() float64 {
 	// number of visits to this node.
 	// Q(s, a) = 1/N(s, a) * Σ(Rᵢ) for i from 1 to N(s, a), where Rᵢ represents the reward
 	// received after the i-th visit to the node, and N(s, a) is the total number of visits.
-	avgReward := n.wins / float64(n.visits)
+	avgReward := n.wins / n.visits
 
 	// If there is no parent (root node), just return the average reward.
 	if n.parent == nil {
